+++ R/tm_g_ae_oview.R
@@ -81,7 +81,8 @@ tm_g_ae_oview <- function(label,
     server = srv_g_ae_oview,
     server_args = list(label = label,
                        dataname = dataname,
-                       add_flag = add_flag),
+                       add_flag = add_flag,
+                       plot_height = plot_height),
     ui = ui_g_ae_oview,
     ui_args = args,
     filters = dataname
@@ -92,7 +93,7 @@ ui_g_ae_oview <- function(id, ...) {
   ns <- NS(id)
   args <- list(...)
   standard_layout(
-    output = white_small_well(plot_decorate_output(id = ns(NULL))),
+    output = white_small_well(plot_decorate_output(id = ns(NULL), plot_height = args$plot_height)),
     encoding = div(
       optionalSelectInput(
         ns("arm_var"),
@@ -151,7 +152,6 @@ ui_g_ae_oview <- function(id, ...) {
       ),
       ui_g_decorate(
         ns(NULL),
-        plot_height = args$plot_height,
         fontsize = args$fontsize,
         titles = "AE Overview",
         footnotes = ""
@@ -169,9 +169,10 @@ srv_g_ae_oview <- function(input,
                            datasets,
                            dataname,
                            label,
-                           add_flag) {
+                           add_flag,
+                           plot_height) {
   init_chunks()
-  font_size <- callModule(srv_g_decorate, id = NULL, plt = plt) # nolint
+  font_size <- callModule(srv_g_decorate, id = NULL, plt = plt, height = plot_height) # nolint
 
   observe({
     req(!is.null(input$diff_ci_method) && !is.null(input$conf_level))